# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Endurance API",
    "description": "Calculations for running, cycling, swimming, and triathlon.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/vdot/calculate": {
      "post": {
        "tags": [
          "VDOT"
        ],
        "summary": "Calculate VDOT from race performance",
        "description": "Calculate VDOT from a race time using Jack Daniels' formula.\n\nVDOT = (VO2max * 100) / (percent_max * oxygen_cost)\nSimplified: VDOT \u2248 (velocity_mps * 60 * 1000) / (3.5 + distance_factor)",
        "operationId": "calculate_vdot_vdot_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VDOTRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VDOTResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/vdot/paces": {
      "get": {
        "tags": [
          "VDOT"
        ],
        "summary": "Get training paces for a VDOT",
        "description": "Return training paces (sec/km) for a given VDOT.",
        "operationId": "get_paces_vdot_paces_get",
        "parameters": [
          {
            "name": "vdot",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 85,
              "minimum": 30,
              "title": "Vdot"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/vdot/race-times": {
      "get": {
        "tags": [
          "VDOT"
        ],
        "summary": "Predict race times for a VDOT",
        "description": "Predict race times for common distances at a given VDOT.",
        "operationId": "predict_times_vdot_race_times_get",
        "parameters": [
          {
            "name": "vdot",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 85,
              "minimum": 30,
              "title": "Vdot"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "VDOTRequest": {
        "properties": {
          "distance_m": {
            "type": "integer",
            "minimum": 100.0,
            "title": "Distance M",
            "description": "Race distance in meters"
          },
          "time_seconds": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Time Seconds",
            "description": "Finish time in seconds"
          }
        },
        "type": "object",
        "required": [
          "distance_m",
          "time_seconds"
        ],
        "title": "VDOTRequest"
      },
      "VDOTResponse": {
        "properties": {
          "vdot": {
            "type": "number",
            "title": "Vdot",
            "description": "Calculated VDOT score"
          },
          "vo2max_est": {
            "type": "number",
            "title": "Vo2Max Est",
            "description": "Estimated VO2max (ml/kg/min)"
          },
          "training_paces": {
            "additionalProperties": true,
            "type": "object",
            "title": "Training Paces",
            "description": "Training paces (sec/km)"
          }
        },
        "type": "object",
        "required": [
          "vdot",
          "vo2max_est",
          "training_paces"
        ],
        "title": "VDOTResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}